Use int64 for timestamps
Summary:
From Hyves:

There are a couple of functions that use int(32) for timestamps, so for
example the following code doesn't print what would be expected:

Test Plan:
<?php
echo date(DateTime::ATOM, mktime(0, 0, 0, 1, 1, 2100)), PHP_EOL;

Reviewed By: qixin
Reviewers: hzhao, qixin, mwilliams
CC: qixin
Revert Plan:
OK

Differential Revision: 224395

+++ src/runtime/base/time/timestamp.cpp
@@ -23,7 +23,7 @@ namespace HPHP {
 ///////////////////////////////////////////////////////////////////////////////
 // creation
 
-int TimeStamp::Current() {
+int64 TimeStamp::Current() {
   return time(0);
 }
 
@@ -58,7 +58,7 @@ String TimeStamp::CurrentMicroTime() {
   return String(ret, CopyString);
 }
 
-int TimeStamp::Get(bool &error, int hou, int min, int sec, int mon, int day,
+int64 TimeStamp::Get(bool &error, int hou, int min, int sec, int mon, int day,
                    int yea, bool gmt) {
   DateTime dt(Current());
   if (gmt) {
